all: join the bundled x/tools files into one
The split mirrored `go tool bundle`'s one-file-per-package output, which no
longer reflects how these files are maintained: the pruning and patching are
applied by an agent following AGENTS.md rather than regenerated, so a single
file with one version marker and one import block is simpler to re-sync.
A NOTE(garble) comment introduces each package's code.
