+++ AGENTS.md
@@ -13,7 +13,7 @@ architecture (toolexec model, caching, terminology) and `README.md` for design g
   (`cache_*.go`), runtime/linker patching (`runtime_patch.go`).
 - `internal/`: `literals` (string obfuscation), `ctrlflow`, `linker`, `ssa2ast`,
   `asthelper`.
-- `bundled_*.go`: hand-patched copies of x/tools packages; see below before touching.
+- `bundled_xtools.go`: patched copies of x/tools packages; see below before touching.
 - `go_std_tables.go` is generated by `scripts/gen_go_std_tables.go` (`go generate`).
 
 ## Testing
@@ -35,19 +35,18 @@ architecture (toolexec model, caching, terminology) and `README.md` for design g
 
 ## Bundled x/tools packages
 
-`bundled_typeutil.go` and `bundled_typeparams.go` are modified copies of x/tools
-packages, needed so that `typeutil.Hasher` does not hash struct field tags, given
-that tags do not affect type identity. typeutil imports the internal `typeparams`
-package, so that is bundled as well, with the import replaced by the bundled
-`typeparams_` names.
+`bundled_xtools.go` holds modified copies of x/tools packages, needed so that
+`typeutil.Hasher` does not hash struct field tags, given that tags do not affect type
+identity. typeutil imports the internal `typeparams` package, so that is bundled too,
+with the import replaced by the bundled `typeparams_` names. Each package's code is
+prefixed with its name and introduced by a `NOTE(garble)` comment naming its origin.
 
-Our patches are marked `// NOTE(garble)` and are done by hand. We also drop the
-`//go:generate` lines that `bundle` adds, as re-running it would undo our edits, and
-any API which is not reachable from `typeutil_Hasher`. Line 1 of each file records
-the x/tools version it was bundled from.
+Our patches are marked `// NOTE(garble)`. They are applied by an agent following the
+steps below, not regenerated by a tool. We also drop any API which is not reachable
+from `typeutil_Hasher`. Line 1 records the x/tools version the file was bundled from.
 
-Bumping `golang.org/x/tools` requires re-syncing both files. Never overwrite them
-with raw `go tool bundle` output, as the pruning and patches would have to be redone.
+Bumping `golang.org/x/tools` requires re-syncing the file. Never overwrite it with
+raw `go tool bundle` output, as the pruning and patches would have to be redone.
 Instead:
 
 * `go build -o /tmp/bundle golang.org/x/tools/cmd/bundle` after the bump; use this
@@ -55,9 +54,9 @@ Instead:
 * bundle both packages at the old version via a throwaway module requiring it and
   blank-importing them, e.g. `/tmp/bundle -o /tmp/old_typeutil.go
   golang.org/x/tools/go/types/typeutil`, and at the new version from this repository
-* diff old against new to see what upstream changed; hand-apply only the hunks that
-  touch code we still keep
-* diff each `bundled_*.go` against the new bundle to confirm the only differences are
+* diff old against new to see what upstream changed; apply only the hunks that touch
+  code we still keep
+* diff `bundled_xtools.go` against the new bundles to confirm the only differences are
   the removed API and our `NOTE(garble)` patches, then bump the version in line 1
 * `go build ./...` and check `staticcheck` reports no unused code
